Update StoreChangeObserver; move selection and clipboard to editor folder
