+++ src/renderer/components/Word.tsx
@@ -9,7 +9,7 @@ import {
 import { MousePosition } from '@react-hook/mouse-position';
 import { pointIsInsideRect } from 'renderer/util';
 import colors from '../colors';
-import { handleSelectWord } from '../selection';
+import { handleSelectWord } from '../editor/selection';
 import { DragState } from './WordDragManager';
 
 const makeWordInner = (isDragActive: boolean) =>
